# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SLSEA Solar Generation API",
    "version": "1.0.0",
    "description": "RESTful API for Sri Lanka Sustainable Energy Authority solar generation telemetry, geographic hierarchy, and operational analytics."
  },
  "servers": [
    {
      "url": "https://slsea-solar-api-git-dev-diluka-theekshana-s-projects.vercel.app",
      "description": "Production Vercel Serverless"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "UserAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "UserAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-User-Id",
        "description": "User ID or Username for RBAC: 1 / national_admin, 2 / provincial_officer, 3 / district_operator"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key for meter telemetry device ingestion (e.g. key_inst_1)"
      }
    },
    "schemas": {
      "Province": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Western Province"
          },
          "code": {
            "type": "string",
            "example": "WP"
          }
        }
      },
      "District": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Colombo"
          },
          "code": {
            "type": "string",
            "example": "CM"
          },
          "province_id": {
            "type": "integer",
            "example": 1
          }
        }
      },
      "GridSubstation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Colombo Central Grid Substation"
          },
          "capacity_mva": {
            "type": "number",
            "example": 100
          },
          "district_id": {
            "type": "integer",
            "example": 1
          }
        }
      },
      "SolarInstallation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Solar Installation 001"
          },
          "meter_id": {
            "type": "string",
            "example": "SLMTR-00001"
          },
          "inverter_id": {
            "type": "string",
            "example": "INV-00001"
          },
          "substation_id": {
            "type": "integer",
            "example": 1
          },
          "latitude": {
            "type": "number",
            "example": 6.939648
          },
          "longitude": {
            "type": "number",
            "example": 79.818451
          },
          "capacity_kw": {
            "type": "number",
            "example": 8
          }
        }
      },
...
      "DistrictSummary": {
        "type": "object",
        "properties": {
          "district_id": {
            "type": "integer",
            "example": 1
          },
          "district_name": {
            "type": "string",
            "example": "Colombo"
          },
          "total_installations": {
            "type": "integer",
            "example": 10
          },
          "current_power_kw": {
            "type": "number",
            "example": 0
          },
          "total_energy_kwh": {
            "type": "number",
            "example": 4034.84
          },
          "peak_power_kw": {
            "type": "number",
            "example": 13.61
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "RESOURCE_NOT_FOUND"
              },
              "message": {
                "type": "string",
                "example": "Province not found"
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/provinces": {
      "get": {
        "summary": "List all provinces",
        "description": "Query all provinces from MongoDB",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Bare array of all provinces",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Province"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Provinces"
        ]
      }
    },
    "/provinces/{id}": {
      "get": {
        "summary": "Get province by ID",
        "description": "Find a province where id matches numeric ID parameter",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Province object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Province"
                }
              }
            }
          },
          "401": {
            "description": "Invalid user identification",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access denied: resource outside jurisdiction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Province not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Provinces"
        ]
      }
    },
    "/provinces/{id}/districts": {
      "get": {
        "summary": "List districts in province",
        "description": "Query all districts where province_id matches numeric ID",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of scoped districts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/District"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Province not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Provinces"
        ]
      }
    },
    "/districts": {
      "get": {
        "summary": "List all districts",
        "description": "Query all districts from MongoDB",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Bare array of all districts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/District"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Districts"
        ]
      }
    },
    "/districts/{id}": {
      "get": {
        "summary": "Get district by ID",
        "description": "Find a district where id matches numeric ID parameter",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "District object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/District"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "District not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Districts"
        ]
      }
    },
    "/districts/{id}/substations": {
      "get": {
        "summary": "List substations in district",
        "description": "Query grid substations where district_id matches numeric ID",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of scoped substations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GridSubstation"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "District not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Substations"
        ]
      }
    },
    "/districts/{id}/summary": {
      "get": {
        "summary": "Get district analytical summary",
        "description": "Upper-band analytical stretch endpoint aggregating current power, total energy, and peak power for a district",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "District analytical summary metrics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistrictSummary"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "District not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Districts"
        ]
      }
    },
    "/substations/{id}/installations": {
      "get": {
        "summary": "List installations in substation",
        "description": "Query solar installations where substation_id matches numeric ID",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of scoped solar installations (api_key excluded)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SolarInstallation"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Substation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Substations"
        ]
      }
    },
...
    "/substations": {
      "get": {
        "summary": "List all grid substations",
        "description": "Query all grid substations across Sri Lanka from MongoDB",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Bare array of all grid substations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GridSubstation"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Substations"
        ]
      }
    },
    "/substations/{id}": {
      "get": {
        "summary": "Get grid substation by ID",
        "description": "Find a grid substation where id matches numeric ID parameter",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Grid substation object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GridSubstation"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Substation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Substations"
        ]
      }
    },
...
  },
...
}